requirements.txt
setuptools
big-fish==0.6.2
cellpose==2.2.2
cloudpickle==2.2.1
dask==2023.9.1
distributed==2023.9.1
fastapi==0.100.0
fpdf==1.7.2
joblib==1.3.2
joypy==0.2.6
keras==2.12.0
matplotlib==3.7.2
matplotlib-scalebar==0.8.1
numpy==1.23.5
opencv-python==4.8.0.74
pandas==1.4.4
Pillow==10.0.0
pycromanager==0.27.5
pyserial==3.5
pysmb==1.2.9.1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.1
torch==2.0.1
tornado==6.3.3
uvicorn==0.23.1
seaborn
